Moved some files to the newly created layers directory, removed $pass_route parameter to Flight::route, to the Route and to the Router::map, added a middleware to do the same as a static function inside FlightLayers class
