+++ flight/LayersStack.php
@@ -22,20 +22,21 @@ class LayersStack {
     }
     
     /**
-    * Dispatch is a repeatable middlewares dispatch function that start the iteration on the current stack of layers before reaching the final route dispatch
+    * Dispatch is a repeatable middlewares dispatch function that start the iteration on the current stack of layers until there are no more
+    *
     * You shouldn't invoke it directly, instead map this method to Flight as "dispatchRoute" with Flight::map( 'dispatchRoute', [ $thisObject, 'dispatch' ] 
-    * overriding the original mapped method Flight::dispatchRoute and let Flight call it for you when it find a compatible route for the request.
-    * As it may found multiple routes compatibile with yours the repeatability of the dispatch function is a must
+    * overriding the original method and let Flight call it for you when it find a compatible route for the request.
+    * As the router may found multiple routes compatibile with yours the repeatability of the dispatch function is a must
     */
     public function dispatch(net\Route $route, array $params) {
-        $iterator = new LayersIterator( $this->layers );
+        $iterator = new layers\LayersIterator( $this->layers );
         $iterator->next($route, $params, \Flight::request(), \Flight::response);
     }
     
     /**
     * RealDispatch is a super thin wrapper over the original Flight::dispatchRoute
     */
-    public function realDispatch(net\Route $route, array $params, net\Request $request, net\Response $response, LayersIterator $iterator) {
+    public function realDispatch(net\Route $route, array $params, net\Request $request, net\Response $response, layers\LayersIterator $iterator) {
         \Flight::_dispatchRoute($route, $params);
     }
