+++ flight/layers/LayersIterator.php
@@ -1,6 +1,10 @@
 <?php
 
-namespace flight;
+namespace flight\layers;
+
+use flight\net\Route;
+use flight\net\Request;
+use flight\net\Response;
 
 /**
 * LayersIterator iterate over the middlewares stack and call the next layer until it abort or reach the final layer (the route layer)
@@ -20,13 +24,13 @@ class LayersIterator {
 
     /**
     * Dispatch the next layer of the stack
-    * @param net\Route $route route found by the router
+    * @param Route $route route found by the router
     * @param array $params list of params for the route
-    * @param net\Request $request Flight request object
-    * @param net\Response $response Flight response object
+    * @param Request $request Flight request object
+    * @param Response $response Flight response object
     * @return void
     */
-    public function next(net\Route $route, array $params, net\Request $request, net\Response $response) {
+    public function next(Route $route, array $params, Request $request, Response $response) {
         if ( $this->cursor > -1 ) {
             $callable = $this->layers[ $this->cursor ];
             $this->cursor--;
